+++ app/src/main/java/com/sdhong/jonbeowin/feature/jonbeocount/JonbeoCountListAdapter.kt
@@ -1,4 +1,4 @@
-package com.sdhong.jonbeowin.view
+package com.sdhong.jonbeowin.feature.jonbeocount
 
 import android.view.LayoutInflater
 import android.view.View
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.sdhong.jonbeowin.databinding.ItemAssetBinding
 import com.sdhong.jonbeowin.local.model.Asset
-import com.sdhong.jonbeowin.view.uistate.AssetUiState
+import com.sdhong.jonbeowin.feature.jonbeocount.uistate.AssetUiState
 
 class JonbeoCountListAdapter(
     private val onAssetItemClick: (asset: Asset) -> Unit
