+++ app/src/main/res/layout/fragment_jonbeo_count.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.JonbeoCountFragment">
+    tools:context=".feature.jonbeocount.JonbeoCountFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbarJonbeocount"
